requirements.txt
charset-normalizer==3.4.4
imageio==2.37.2
nuitka==4.0
numpy==2.4.2
pillow==12.2.0
reportlab==4.4.9
zstandard==0.25.0
